requirements.txt
aenum==2.2.3
altgraph==0.17
asgiref==3.2.10
astroid==2.3.3
backcall==0.1.0
cffi==1.13.2
Click==7.0
colorama==0.4.3
contextlib2==0.6.0.post1
cryptography==2.8
cycler==0.10.0
debug==0.3.2
decorator==4.4.1
Django==3.1
django-progressive-web-app==0.1.1
django-pwa==1.0.10
ecdsa==0.15
esptool==2.8
Flask==1.1.2
future==0.18.2
gunicorn==20.0.4
httplib2==0.17.0
ipdb==0.12.3
ipython==7.12.0
ipython-genutils==0.2.0
iso8601==0.1.12
isort==4.3.21
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.2
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.1.2
mccabe==0.6.1
mutablerecords==0.4.1
numpy==1.18.1
oauth2client==1.5.2
pandas==1.0.0
parso==0.6.1
pefile==2019.4.18
pickleshare==0.7.5
plotly==4.5.4
prompt-toolkit==3.0.3
protobuf==3.11.2
py2exe==0.9.2.2
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
Pygments==2.5.2
PyInstaller==3.6
pylint==2.4.4
pyOpenSSL==17.5.0
pyparsing==2.4.6
PyQt5==5.14.1
PyQt5-sip==12.7.0
pyqt5-tools==5.13.2.1.6rc1
PyQtChart==5.14.0
pyqtgraph==0.10.0
pyserial==3.4
pyshortcuts==1.7.1
PySide2==5.14.1
python-can==3.3.2
python-dateutil==2.8.1
python-dotenv==0.10.5
PythonQwt==0.5.5
pytz==2019.3
pywin32==227
pywin32-ctypes==0.2.0
PyYAML==3.13
retrying==1.3.3
rsa==4.0
see==1.4.1
shiboken2==5.14.1
six==1.14.0
sockjs-tornado==1.0.6
sqlparse==0.3.1
stopwatch.py==1.0.1
tornado==4.5.3
traitlets==4.3.3
wcwidth==0.1.8
Werkzeug==1.0.1
whitenoise==5.2.0
windows-curses==2.1.0
wrapt==1.11.2
